backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
supabase>=2.0.0
aiohttp>=3.8.0
python-multipart>=0.0.5
python-dotenv>=1.0.0
